# docs/swagger.yaml
basePath: /api
definitions:
...
  controllers.DeviceRequest:
    type: object
...
  controllers.ErrorResponse:
    properties:
      code:
        example: 401
        type: integer
      data: {}
      message:
        example: Invalid username or password
        type: string
    type: object
...
  models.AccessLog:
    properties:
      device:
        allOf:
        - $ref: '#/definitions/models.Device'
        description: Relations
      device_id:
        type: integer
      id:
        type: integer
      method:
        $ref: '#/definitions/models.AccessMethod'
      resident:
        $ref: '#/definitions/models.Resident'
      resident_id:
        type: integer
      result:
        $ref: '#/definitions/models.AccessResult'
      timestamp:
        type: string
    type: object
  models.AccessMethod:
    enum:
    - remote
    - code
    - face
    - fingerprint
    type: string
    x-enum-varnames:
    - AccessMethodRemote
    - AccessMethodCode
    - AccessMethodFace
    - AccessMethodFingerprint
  models.AccessResult:
    enum:
    - success
    - failure
    type: string
    x-enum-varnames:
    - AccessResultSuccess
    - AccessResultFailure
  models.CallRecord:
    properties:
      call_status:
        $ref: '#/definitions/models.CallStatus'
      device:
        allOf:
        - $ref: '#/definitions/models.Device'
        description: Relations
      device_id:
        type: integer
      duration:
        description: in seconds
        type: integer
      id:
        type: integer
      resident:
        $ref: '#/definitions/models.Resident'
      resident_id:
        type: integer
      timestamp:
        type: string
    type: object
  models.CallStatus:
    enum:
    - answered
    - missed
    - timeout
    type: string
    x-enum-varnames:
    - CallStatusAnswered
    - CallStatusMissed
    - CallStatusTimeout
  models.Device:
    properties:
      access_logs:
        items:
          $ref: '#/definitions/models.AccessLog'
        type: array
      call_records:
        items:
          $ref: '#/definitions/models.CallRecord'
        type: array
      created_at:
        type: string
      emergency_logs:
        items:
          $ref: '#/definitions/models.EmergencyLog'
        type: array
      id:
        type: integer
      location:
        type: string
      name:
        type: string
      resident:
        $ref: '#/definitions/models.Resident'
      serial_number:
        type: string
      staff:
        description: Relations - 多对多关系
        items:
          $ref: '#/definitions/models.PropertyStaff'
        type: array
      status:
        $ref: '#/definitions/models.DeviceStatus'
      updated_at:
        type: string
    type: object
  models.DeviceStatus:
    enum:
    - online
    - offline
    - fault
    type: string
    x-enum-varnames:
    - DeviceStatusOnline
    - DeviceStatusOffline
    - DeviceStatusFault
  models.EmergencyLog:
    properties:
      device:
        allOf:
        - $ref: '#/definitions/models.Device'
        description: Relations
      device_id:
        type: integer
      id:
        type: integer
      resident:
        $ref: '#/definitions/models.Resident'
      resident_id:
        type: integer
      resolved_at:
        description: 可空字段
        type: string
      status:
        $ref: '#/definitions/models.EmergencyStatus'
      triggered_at:
        type: string
    type: object
  models.EmergencyStatus:
    enum:
    - pending
    - responded
    - escalated
    - resolved
    type: string
    x-enum-varnames:
    - EmergencyStatusPending
    - EmergencyStatusResponded
    - EmergencyStatusEscalated
    - EmergencyStatusResolved
  models.PropertyStaff:
    properties:
      created_at:
        type: string
      devices:
        description: 关联关系 - 使用多对多关系替代直接关联
        items:
          $ref: '#/definitions/models.Device'
        type: array
      id:
        type: integer
      phone:
        type: string
      position:
        type: string
      property_name:
        type: string
      remark:
        type: string
      role:
        description: manager, staff, etc.
        type: string
      status:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.Resident:
    properties:
      access_logs:
        items:
          $ref: '#/definitions/models.AccessLog'
        type: array
      call_records:
        items:
          $ref: '#/definitions/models.CallRecord'
        type: array
      created_at:
        type: string
      device:
        allOf:
        - $ref: '#/definitions/models.Device'
        description: Relations
      device_id:
        description: One-to-one relationship with Device
        type: integer
      email:
        type: string
      emergency_logs:
        items:
          $ref: '#/definitions/models.EmergencyLog'
        type: array
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:20033
info:
  contact:
    email: support@yourcompany.com
    name: API Support
    url: http://www.yourcompany.com/support
  description: A comprehensive door access management system with video calling capabilities
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ILock HTTP Service API
  version: "1.0"
paths:
...
  /devices:
    get:
      consumes:
      - application/json
      description: 获取所有设备的列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Device'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取所有设备
      tags:
      - device
    post:
      consumes:
      - application/json
      description: 创建一个新的门禁设备
      parameters:
      - description: 设备信息
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/controllers.DeviceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Device'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建新设备
      tags:
      - device
  /devices/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除设备
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除设备
      tags:
      - device
    get:
      consumes:
      - application/json
      description: 根据ID获取设备信息
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Device'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取单个设备
      tags:
      - device
    put:
      consumes:
      - application/json
      description: 根据ID更新设备信息
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: string
      - description: 设备信息
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/controllers.DeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Device'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新设备信息
      tags:
      - device
  /devices/{id}/status:
    get:
      consumes:
      - application/json
      description: Get the current status of a device by ID
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Device Status
      tags:
      - device
...
securityDefinitions:
  BearerAuth:
    description: 'Enter the token with the `Bearer: ` prefix'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"